pyproject.toml
[tool.poetry]
name = "medical-diagnostic1"
version = "0.1.0"
description = ""
authors = ["Alexandr Zibinin <zibinin.chel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
coverage = "^7.6.1"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
